first commit
Here the functionality of creating new user , and creating room for the game, and joining to the room created by other user using room name are been developed and the startgame button will distribute the cards randomly and create the carlist for each player in the room. and displaying the cards each user has in his screen.